fix/improvements

+++ source/bxgeomtools/src/box.cc
@@ -142,7 +142,7 @@ namespace geomtools {
   {
     double s = 0.0;
     int mask = mask_;
-    if (mask_ == ALL_SURFACES) mask = FACE_ALL;
+    if (mask_ == (int) ALL_SURFACES) mask = FACE_ALL;
 
     if (mask & FACE_BACK) 
       {
@@ -250,7 +250,7 @@ namespace geomtools {
     if (skin_ > USING_PROPER_SKIN) skin = 2 * skin_;
 
     int mask = mask_;
-    if (mask_ == ALL_SURFACES) mask = FACE_ALL;
+    if (mask_ == (int) ALL_SURFACES) mask = FACE_ALL;
 
     double hskin = 0.5 * skin;
     if (mask & FACE_BACK)
